Fixed the remaining integer overflows in the df-occ and cd-occ memory demand estimation. long long int was chosen deliberately to stay exact as long as possible. These calculations are furthermore not time critical, thus the penalty of 64 bit integer arithmetics is negligible.

+++ psi4/src/psi4/dfocc/manager.cc
@@ -621,7 +621,7 @@ void DFOCC::ccsd_manager() {
 
         // Mem for amplitudes
         cost_ampAA = 0.0;
-        cost_ampAA = naocc2AA * nvir2AA;
+        cost_ampAA = (long long int)naocc2AA * (long long int)nvir2AA;
         cost_ampAA /= 1024.0 * 1024.0;
         cost_ampAA *= sizeof(double);
         cost_3amp = 3.0 * cost_ampAA;
